feat: lay foundation for context driven filters and operators
