+++ packages/sanity/src/core/studio/components/navbar/search/components/filters/common/FilterIcon.tsx
@@ -2,13 +2,18 @@ import {UnknownIcon} from '@sanity/icons'
 import React, {createElement} from 'react'
 import {getFilterDefinition} from '../../../definitions/filters'
 import type {SearchFilter} from '../../../types'
+import {useSearchState} from '../../../contexts/search/useSearchState'
 
 interface FilterIconProps {
   filter: SearchFilter
 }
 
 export function FilterIcon({filter}: FilterIconProps) {
-  const icon = getFilterDefinition(filter.filterType)?.icon
+  const {
+    state: {definitions},
+  } = useSearchState()
+
+  const icon = getFilterDefinition(definitions.filters, filter.filterType)?.icon
   if (icon) {
     return createElement(icon)
   }
