+++ packages/sanity/src/core/studio/components/navbar/search/definitions/filters.ts
@@ -1,31 +1,17 @@
-import {
-  CalendarIcon,
-  CheckmarkCircleIcon,
-  DocumentIcon,
-  ImageIcon,
-  LinkIcon,
-  PinIcon,
-  UlistIcon,
-} from '@sanity/icons'
 import {IntrinsicTypeName} from '@sanity/types'
 import {ComponentType} from 'react'
-import {NumberIcon} from '../components/filters/icons/NumberIcon'
-import {StringIcon} from '../components/filters/icons/StringIcon'
-import {SearchOperatorType} from './operators'
+import {SearchOperatorType} from './operators/defaultOperators'
 
-export type FilterDefinitionType = SearchFilterDefinition['type']
-export type SupportedFilterDefinitionFieldType = NonNullable<SearchFilterDefinition['fieldType']>
-
-export interface SearchFilterDefinition {
+export interface SearchFilterDefinition<TOperator = SearchOperatorType> {
   fieldType: IntrinsicTypeName | 'email' | null
   icon: ComponentType
-  initialOperator: SearchOperatorType
+  initialOperator: TOperator
   operators: (
     | {
         type: 'divider'
       }
     | {
-        name: SearchOperatorType
+        name: TOperator
         type: 'item'
       }
   )[]
@@ -33,237 +19,22 @@ export interface SearchFilterDefinition {
   type: string
 }
 
-export const FILTER_DEFINITIONS: SearchFilterDefinition[] = [
-  {
-    fieldType: 'array',
-    icon: UlistIcon,
-    initialOperator: 'arrayCountEqual',
-    operators: [
-      {name: 'arrayCountEqual', type: 'item'},
-      {name: 'arrayCountNotEqual', type: 'item'},
-      {type: 'divider'},
-      {name: 'arrayCountGt', type: 'item'},
-      {name: 'arrayCountGte', type: 'item'},
-      {name: 'arrayCountLt', type: 'item'},
-      {name: 'arrayCountLte', type: 'item'},
-      {type: 'divider'},
-      {name: 'defined', type: 'item'},
-      {name: 'notDefined', type: 'item'},
-    ],
-    title: 'Array',
-    type: 'array',
-  },
-  {
-    fieldType: 'boolean',
-    icon: CheckmarkCircleIcon,
-    initialOperator: 'booleanEqual',
-    operators: [
-      {name: 'booleanEqual', type: 'item'},
-      {type: 'divider'},
-      {name: 'notDefined', type: 'item'},
-    ],
-    title: 'Boolean',
-    type: 'boolean',
-  },
-  {
-    fieldType: 'date',
-    icon: CalendarIcon,
-    initialOperator: 'dateLast',
-    operators: [
-      {name: 'dateLast', type: 'item'},
-      {name: 'dateAfter', type: 'item'},
-      {name: 'dateBefore', type: 'item'},
-      {type: 'divider'},
-      {name: 'dateRange', type: 'item'},
-      {type: 'divider'},
-      {name: 'dateEqual', type: 'item'},
-      {name: 'dateNotEqual', type: 'item'},
-      {type: 'divider'},
-      {name: 'defined', type: 'item'},
-      {name: 'notDefined', type: 'item'},
-    ],
-    title: 'Date',
-    type: 'date',
-  },
-  {
-    fieldType: 'datetime',
-    icon: CalendarIcon,
-    initialOperator: 'dateLast',
-    operators: [
-      {name: 'dateLast', type: 'item'},
-      {name: 'dateAfter', type: 'item'},
-      {name: 'dateBefore', type: 'item'},
-      {type: 'divider'},
-      {name: 'dateRange', type: 'item'},
-      {type: 'divider'},
-      {name: 'dateEqual', type: 'item'},
-      {name: 'dateNotEqual', type: 'item'},
-      {type: 'divider'},
-      {name: 'defined', type: 'item'},
-      {name: 'notDefined', type: 'item'},
-    ],
-    title: 'Datetime',
-    type: 'datetime',
-  },
-  {
-    fieldType: 'file',
-    icon: DocumentIcon,
-    initialOperator: 'defined',
-    operators: [
-      {name: 'defined', type: 'item'},
-      {name: 'notDefined', type: 'item'},
-    ],
-    title: 'File',
-    type: 'file',
-  },
-  {
-    fieldType: 'geopoint',
-    icon: PinIcon,
-    initialOperator: 'defined',
-    operators: [
-      {name: 'defined', type: 'item'},
-      {name: 'notDefined', type: 'item'},
-    ],
-    title: 'Geopoint',
-    type: 'geopoint',
-  },
-  {
-    fieldType: 'image',
-    icon: ImageIcon,
-    initialOperator: 'defined',
-    operators: [
-      {name: 'defined', type: 'item'},
-      {name: 'notDefined', type: 'item'},
-    ],
-    title: 'Image',
-    type: 'image',
-  },
-  {
-    fieldType: 'number',
-    icon: NumberIcon,
-    initialOperator: 'numberEqual',
-    operators: [
-      {name: 'numberEqual', type: 'item'},
-      {name: 'numberNotEqual', type: 'item'},
-      {type: 'divider'},
-      {name: 'numberGt', type: 'item'},
-      {name: 'numberGte', type: 'item'},
-      {name: 'numberLt', type: 'item'},
-      {name: 'numberLte', type: 'item'},
-      {type: 'divider'},
-      {name: 'numberRange', type: 'item'},
-      {type: 'divider'},
-      {name: 'defined', type: 'item'},
-      {name: 'notDefined', type: 'item'},
-    ],
-    title: 'Number',
-    type: 'number',
-  },
-  {
-    fieldType: 'reference',
-    icon: LinkIcon,
-    initialOperator: 'referenceEqual',
-    operators: [
-      {name: 'referenceEqual', type: 'item'},
-      {type: 'divider'},
-      {name: 'defined', type: 'item'},
-      {name: 'notDefined', type: 'item'},
-    ],
-    title: 'Reference',
-    type: 'reference',
-  },
-  {
-    fieldType: null,
-    icon: LinkIcon,
-    initialOperator: 'references',
-    operators: [{name: 'references', type: 'item'}],
-    title: 'Referenced document',
-    type: 'references',
-  },
-  {
-    fieldType: 'slug',
-    icon: StringIcon,
-    initialOperator: 'stringMatches',
-    operators: [
-      {name: 'stringMatches', type: 'item'},
-      {name: 'stringNotMatches', type: 'item'},
-      {type: 'divider'},
-      {name: 'stringEqual', type: 'item'},
-      {name: 'stringNotEqual', type: 'item'},
-      {type: 'divider'},
-      {name: 'defined', type: 'item'},
-      {name: 'notDefined', type: 'item'},
-    ],
-    title: 'Slug',
-    type: 'slug',
-  },
-  {
-    fieldType: 'string',
-    icon: StringIcon,
-    initialOperator: 'stringMatches',
-    operators: [
-      {name: 'stringMatches', type: 'item'},
-      {name: 'stringNotMatches', type: 'item'},
-      {type: 'divider'},
-      {name: 'stringEqual', type: 'item'},
-      {name: 'stringNotEqual', type: 'item'},
-      {type: 'divider'},
-      {name: 'defined', type: 'item'},
-      {name: 'notDefined', type: 'item'},
-    ],
-    title: 'String',
-    type: 'string',
-  },
-  {
-    fieldType: 'text',
-    icon: StringIcon,
-    initialOperator: 'stringMatches',
-    operators: [
-      {name: 'stringMatches', type: 'item'},
-      {name: 'stringNotMatches', type: 'item'},
-      {type: 'divider'},
-      {name: 'stringEqual', type: 'item'},
-      {name: 'stringNotEqual', type: 'item'},
-      {type: 'divider'},
-      {name: 'defined', type: 'item'},
-      {name: 'notDefined', type: 'item'},
-    ],
-    title: 'Text',
-    type: 'text',
-  },
-  {
-    fieldType: 'url',
-    icon: StringIcon,
-    initialOperator: 'stringMatches',
-    operators: [
-      {name: 'stringMatches', type: 'item'},
-      {name: 'stringNotMatches', type: 'item'},
-      {type: 'divider'},
-      {name: 'stringEqual', type: 'item'},
-      {name: 'stringNotEqual', type: 'item'},
-      {type: 'divider'},
-      {name: 'defined', type: 'item'},
-      {name: 'notDefined', type: 'item'},
-    ],
-    title: 'URL',
-    type: 'url',
-  },
-]
-
 export function getFilterDefinition(
-  filterType: FilterDefinitionType
+  definitions: SearchFilterDefinition[],
+  filterType: string
 ): SearchFilterDefinition | undefined {
-  return FILTER_DEFINITIONS.find((filter) => filter.type === filterType)
+  return definitions.find((filter) => filter.type === filterType)
 }
 
 export function getFilterDefinitionInitialOperator(
-  filterName: FilterDefinitionType
+  definitions: SearchFilterDefinition[],
+  filterName: string
 ): SearchFilterDefinition['initialOperator'] | undefined {
-  return getFilterDefinition(filterName)?.initialOperator
+  return getFilterDefinition(definitions, filterName)?.initialOperator
 }
 
-export function getSupportedFieldTypes(): string[] {
-  return FILTER_DEFINITIONS.reduce<string[]>((acc, val) => {
+export function getSupportedFieldTypes(definitions: SearchFilterDefinition[]): string[] {
+  return definitions.reduce<string[]>((acc, val) => {
     if (val?.fieldType) {
       acc.push(val.fieldType)
     }
