+++ packages/sanity/src/core/studio/components/navbar/search/definitions/operators/booleanOperators.ts
@@ -1,11 +1,12 @@
 import {FieldInputBoolean} from '../../components/filters/filter/inputTypes/Boolean'
 import {toJSON} from './operatorUtils'
-import {defineSearchOperator} from './operatorTypes'
+import {defineSearchOperator, SearchOperatorParams} from './operatorTypes'
 
 export const booleanOperators = {
   booleanEqual: defineSearchOperator({
     buttonLabel: 'is',
-    fn: ({fieldPath, value}) => (value && fieldPath ? `${fieldPath} == ${toJSON(value)}` : null),
+    fn: ({fieldPath, value}: SearchOperatorParams<boolean>) =>
+      value && fieldPath ? `${fieldPath} == ${toJSON(value)}` : null,
     initialValue: true,
     inputComponent: FieldInputBoolean,
     label: 'is',
