+++ packages/sanity/src/core/studio/components/navbar/search/definitions/operators/index.ts
@@ -1,35 +1,22 @@
-import {arrayOperators} from './arrayOperators'
-import {stringOperators} from './stringOperators'
-import {dateOperators} from './dateOperators'
-import {referenceOperators} from './referenceOperators'
-import {numberOperators} from './numberOperators'
-import {definedOperators} from './defineOperators'
-import {booleanOperators} from './booleanOperators'
-import {assetOperators} from './assetOperators'
+import {SearchOperatorBase, SearchOperatorInput, SearchOperatorParams} from './operatorTypes'
 
-const OPERATORS = {
-  ...arrayOperators,
-  ...assetOperators,
-  ...booleanOperators,
-  ...definedOperators,
-  ...dateOperators,
-  ...numberOperators,
-  ...referenceOperators,
-  ...stringOperators,
-} as const
-
-type Operators = typeof OPERATORS
-export type SearchOperatorType = keyof Operators
-export type SearchOperator = Operators[SearchOperatorType]
-
-export function getOperator<T extends SearchOperatorType>(
-  operatorType?: T
-): Operators[T] | undefined {
-  return operatorType ? OPERATORS[operatorType] : undefined
+export interface SearchOperator extends SearchOperatorBase {
+  fn: (params: SearchOperatorParams<any>) => string | null
+  initialValue?: any
+  inputComponent?: SearchOperatorInput<any>
+  type: string
 }
 
-export function getOperatorInitialValue<T extends SearchOperatorType>(
-  operatorType: T
-): Operators[T]['initialValue'] | undefined {
-  return getOperator(operatorType)?.initialValue
+export function getOperator(
+  operators: SearchOperator[],
+  operatorType?: string
+): SearchOperator | undefined {
+  return operatorType ? operators.find((f) => f.type === operatorType) : undefined
+}
+
+export function getOperatorInitialValue(
+  operators: SearchOperator[],
+  operatorType?: string
+): SearchOperator['initialValue'] | undefined {
+  return getOperator(operators, operatorType)?.initialValue
 }
